Add OwnerReference to created network policies
Set the OwnerReference on the NetworkPolicy objects created by the
lifecycle-agent. The owner is determined by looking up the current pod's
owner, which is typically the lifecycle-agent-controller-manager
replicaset.

This change ensures that when the lifecycle-agent is uninstalled, the
created NetworkPolicy resources are automatically garbage collected by
Kubernetes, preventing orphaned resources in the cluster.

```
 -n openshift-lifecycle-agent get networkpolicies lca-prep-stateroot-setup -o yaml
apiVersion: networking.k8s.io/v1
kind: NetworkPolicy
metadata:
  creationTimestamp: "2025-07-18T09:55:53Z"
  generation: 1
  name: lca-prep-stateroot-setup
  namespace: openshift-lifecycle-agent
  ownerReferences:
  - apiVersion: apps/v1
    controller: true
    kind: ReplicaSet
    name: lifecycle-agent-controller-manager-c97b48587
    uid: 81976453-9cfd-4ddd-8783-bb1ea5246619
 ```

Signed-off-by: Konstantinos Karampogias <karampok@gmail.com>